Added permission checks in controllers

+++ api/src/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Controllers;
 
 use App\Data\Repositories\CategoryRepository;
+use App\Data\Repositories\PostRepository;
 use Core\Auth\Attributes\Anonymous;
 use Core\Auth\Attributes\Authorize;
 use Core\Auth\Attributes\Requires;
@@ -18,10 +19,13 @@ use Core\Http\Response;
 #[Route("categories")]
 class CategoryController extends Controller
 {
+    private PostRepository $postRepository;
     private CategoryRepository $categoryRepository;
+
     public function __construct()
     {
         $this->categoryRepository = new CategoryRepository();
+        $this->postRepository = new PostRepository();
     }
 
     #[Anonymous]
@@ -84,7 +88,7 @@ class CategoryController extends Controller
         $categoryId = (int)$request->getRouteParam(0);
         $title = $request->getBody()["title"];
 
-        $this->categoryRepository->UpdateCategory($categoryId ,$title);
+        $this->categoryRepository->UpdateCategory($categoryId, $title);
 
         return $this->json(
             $this->ok()
@@ -105,12 +109,28 @@ class CategoryController extends Controller
         );
     }
 
-    #[Anonymous]
+    #[Authorize]
     #[Post("post/{id}")]
     public function AddPostCategory(Request $request): Response
     {
+        $userId = $this->UserClaim("id");
+        $userRole = $this->UserClaim("role");
         $postId = (int)$request->getRouteParam(0);
         $categoryId = (int)$request->getQueryParams()["categoryid"];
+
+        $res = $this->postRepository->GetPost($userId, $postId);
+
+        if ($res == null) {
+            return $this->json(
+                $this->notFound("Post not found")
+            );
+        }
+
+        if ($res->UserId != $userId && $userRole != "Administrator" && $userRole != "Moderator") {
+            return $this->json(
+                $this->badRequest("Permission denied")
+            );
+        }
 
         $this->categoryRepository->AddPostCategory($postId, $categoryId);
 
@@ -123,8 +143,24 @@ class CategoryController extends Controller
     #[Delete("post/{id}")]
     public function RemovePostCategory(Request $request): Response
     {
+        $userId = $this->UserClaim("id");
+        $userRole = $this->UserClaim("role");
         $postId = (int)$request->getRouteParam(0);
         $categoryId = (int)$request->getQueryParams()["categoryid"];
+
+        $res = $this->postRepository->GetPost($userId, $postId);
+
+        if ($res == null) {
+            return $this->json(
+                $this->notFound("Post not found")
+            );
+        }
+
+        if ($res->UserId != $userId && $userRole != "Administrator" && $userRole != "Moderator") {
+            return $this->json(
+                $this->badRequest("Permission denied")
+            );
+        }
 
         $this->categoryRepository->RemovePostCategory($postId, $categoryId);
